PRO-3820: refactor code to use a single persistence client
