+++ src/test/java/uk/gov/hmcts/probate/services/invitation/PersistenceClientTest.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.probate.services.invitation.model.InviteData;
+import uk.gov.hmcts.probate.services.persistence.PersistenceClient;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
